Move bucket factory to proton test library, it is only used by unit tests.
