+++ searchcore/src/tests/proton/docsummary/docsummary.cpp
@@ -7,7 +7,7 @@
 #include <vespa/eval/tensor/tensor_factory.h>
 #include <vespa/persistence/spi/test.h>
 #include <vespa/searchcore/proton/attribute/attribute_writer.h>
-#include <vespa/searchcore/proton/common/bucketfactory.h>
+#include <vespa/searchcore/proton/test/bucketfactory.h>
 #include <vespa/searchcore/proton/docsummary/docsumcontext.h>
 #include <vespa/searchcore/proton/docsummary/documentstoreadapter.h>
 #include <vespa/searchcore/proton/docsummary/summarymanager.h>
