+++ searchcore/src/tests/proton/server/feedstates_test.cpp
@@ -8,7 +8,7 @@ LOG_SETUP("feedstates_test");
 #include <vespa/document/base/testdocrepo.h>
 #include <vespa/document/bucket/bucketid.h>
 #include <vespa/document/repo/documenttyperepo.h>
-#include <vespa/searchcore/proton/common/bucketfactory.h>
+#include <vespa/searchcore/proton/test/bucketfactory.h>
 #include <vespa/searchcore/proton/server/feedstates.h>
 #include <vespa/searchcore/proton/server/ireplayconfig.h>
 #include <vespa/searchcore/proton/server/memoryconfigstore.h>
